refactor(cpp): split tts methods
