+++ android/src/main/jni.cpp
@@ -816,7 +816,9 @@ Java_com_rnllama_LlamaContext_doCompletion(
             guide_tokens_vector[i] = guide_tokens_array[i];
         }
         env->ReleaseIntArrayElements(guide_tokens, guide_tokens_array, 0);
-        llama->setGuideTokens(guide_tokens_vector);
+        if (llama->tts_ctx != nullptr) {
+            llama->tts_ctx->setGuideTokens(guide_tokens_vector);
+        }
     }
 
     // Process image paths if provided
@@ -1177,7 +1179,11 @@ Java_com_rnllama_LlamaContext_doCompletion(
     if (toolCallsSize > 0) {
         putArray(env, result, "tool_calls", toolCalls);
     }
-    putArray(env, result, "audio_tokens", tokensToArray(env, llama, llama->audio_tokens));
+    std::vector<llama_token> audio_tokens;
+    if (llama->tts_ctx != nullptr) {
+        audio_tokens = llama->tts_ctx->audio_tokens;
+    }
+    putArray(env, result, "audio_tokens", tokensToArray(env, llama, audio_tokens));
     putArray(env, result, "completion_probabilities", tokenProbsToMap(env, llama, llama->generated_token_probs));
     putInt(env, result, "tokens_predicted", llama->num_tokens_predicted);
     putInt(env, result, "tokens_evaluated", llama->num_prompt_tokens);
@@ -1674,7 +1680,7 @@ Java_com_rnllama_LlamaContext_getFormattedAudioCompletion(
 
     auto result = createWriteableMap(env);
     try {
-        auto audio_result = llama->getFormattedAudioCompletion(speaker_json_str_chars, text_to_speak_chars);
+        auto audio_result = llama->tts_ctx->getFormattedAudioCompletion(llama, speaker_json_str_chars, text_to_speak_chars);
         putString(env, result, "prompt", audio_result.prompt.c_str());
         if (audio_result.grammar != nullptr) {
             putString(env, result, "grammar", audio_result.grammar);
@@ -1703,7 +1709,7 @@ Java_com_rnllama_LlamaContext_getAudioCompletionGuideTokens(
     UNUSED(thiz);
     auto llama = context_map[(long) context_ptr];
     const char *text_to_speak_chars = env->GetStringUTFChars(text_to_speak, nullptr);
-    std::vector<llama_token> guide_tokens = llama->getAudioCompletionGuideTokens(text_to_speak_chars);
+    std::vector<llama_token> guide_tokens = llama->tts_ctx->getAudioCompletionGuideTokens(llama, text_to_speak_chars);
     env->ReleaseStringUTFChars(text_to_speak, text_to_speak_chars);
     auto result = createWritableArray(env);
     for (const auto &val : guide_tokens) {
@@ -1729,7 +1735,7 @@ Java_com_rnllama_LlamaContext_decodeAudioTokens(
         tokens_vec[i] = tokens_ptr[i];
     }
     env->ReleaseIntArrayElements(tokens, tokens_ptr, 0);
-    std::vector<float> audio = llama->decodeAudioTokens(tokens_vec);
+    std::vector<float> audio = llama->tts_ctx->decodeAudioTokens(llama, tokens_vec);
     auto result = createWritableArray(env);
     for (const auto &val : audio) {
       pushDouble(env, result, (double) val);
