+++ ios/RNLlamaContext.mm
@@ -652,7 +652,9 @@
         for (NSNumber *token_num in guide_tokens_array) {
             guide_tokens.push_back([token_num intValue]);
         }
-        llama->setGuideTokens(guide_tokens);
+        if (llama->tts_ctx != nullptr) {
+            llama->tts_ctx->setGuideTokens(guide_tokens);
+        }
     }
 
     if (!llama->initSampling()) {
@@ -838,9 +840,9 @@
     result[@"stopping_word"] = [NSString stringWithUTF8String:llama->stopping_word.c_str()];
     result[@"tokens_cached"] = @(llama->n_past);
 
-    if (llama->isVocoderEnabled() && !llama->audio_tokens.empty()) {
+    if (llama->isVocoderEnabled() && llama->tts_ctx != nullptr && !llama->tts_ctx->audio_tokens.empty()) {
         NSMutableArray *audioTokens = [[NSMutableArray alloc] init];
-        for (llama_token token : llama->audio_tokens) {
+        for (llama_token token : llama->tts_ctx->audio_tokens) {
             [audioTokens addObject:@(token)];
         }
         result[@"audio_tokens"] = audioTokens;
@@ -1095,7 +1097,7 @@
 - (NSDictionary *)getFormattedAudioCompletion:(NSString *)speakerJsonStr textToSpeak:(NSString *)textToSpeak {
     std::string speakerStr = speakerJsonStr ? [speakerJsonStr UTF8String] : "";
     try {
-        auto audio_result = llama->getFormattedAudioCompletion(speakerStr, [textToSpeak UTF8String]);
+        auto audio_result = llama->tts_ctx->getFormattedAudioCompletion(llama, speakerStr, [textToSpeak UTF8String]);
         return @{
             @"prompt": [NSString stringWithUTF8String:audio_result.prompt.c_str()],
             @"grammar": [NSString stringWithUTF8String:audio_result.grammar]
@@ -1106,7 +1108,7 @@
 }
 
 - (NSArray *)getAudioCompletionGuideTokens:(NSString *)textToSpeak {
-    std::vector<llama_token> guide_tokens = llama->getAudioCompletionGuideTokens([textToSpeak UTF8String]);
+    std::vector<llama_token> guide_tokens = llama->tts_ctx->getAudioCompletionGuideTokens(llama, [textToSpeak UTF8String]);
     NSMutableArray *result = [[NSMutableArray alloc] init];
     for (llama_token token : guide_tokens) {
         [result addObject:@(token)];
@@ -1119,7 +1121,7 @@
     for (NSNumber *token in tokens) {
         token_vector.push_back([token intValue]);
     }
-    std::vector<float> audio_data = llama->decodeAudioTokens(token_vector);
+    std::vector<float> audio_data = llama->tts_ctx->decodeAudioTokens(llama, token_vector);
     NSMutableArray *result = [[NSMutableArray alloc] init];
     for (float sample : audio_data) {
         [result addObject:@(sample)];
